+++ packages/stripe/src/client/stripe/index.ts
@@ -1,0 +1,2 @@
+export { default as initStripeClient } from './initStripeClient'
+export * from './initStripeClient'