+++ packages/stripe/src/client/supabase/index.ts
@@ -1,0 +1,2 @@
+export { default as StripeSupabaseClient } from './StripeSupabaseClient'
+export * from './StripeSupabaseClient'