+++ packages/stripe/src/server/next/index.ts
@@ -1,0 +1,2 @@
+export { default as StripeRouterMiddleware } from './StripeRouterMiddleware'
+export * from './StripeRouterMiddleware'