+++ packages/stripe/src/server/stripe/initStripeNode.ts
@@ -1,8 +1,8 @@
 import Stripe from 'stripe'
-import stripeConfig from '../config/stripeConfig'
+import stripeConfig from '../../config/stripeConfig'
 
-const getStripeNode = (stripeSecretKey = '') =>
-  new Stripe(stripeSecretKey, {
+const initStripeNode = (stripeSecretKey = '') => {
+  const stripeNode = new Stripe(stripeSecretKey, {
     // https://github.com/stripe/stripe-node#configuration
     apiVersion: '2020-08-27',
     // Register this as an official Stripe plugin.
@@ -12,5 +12,7 @@ const getStripeNode = (stripeSecretKey = '') =>
       version: '0.0.1',
     },
   })
+  return stripeNode
+}
 
-export default getStripeNode
+export default initStripeNode