+++ packages/stripe/src/stripe/index.ts
@@ -1,5 +1,0 @@
-export { default as getStripeClient } from './getStripeClient'
-export * from './getStripeClient'
-
-export { default as getStripeNode } from './getStripeNode'
-export * from './getStripeNode'