+++ packages/stripe/src/supabase/index.ts
@@ -1,5 +1,0 @@
-export { default as getStripeSupabaseAdmin } from './getStripeSupabaseAdmin'
-export * from './getStripeSupabaseAdmin'
-
-export { default as StripeSupabaseClient } from './StripeSupabaseClient'
-export * from './StripeSupabaseClient'