+++ packages/stripe/CHANGELOG.md
@@ -1,5 +1,89 @@
 # Changelog
 
+## 0.0.19
+
+### Patch Changes
+
+- Update stripe middleware
+
+## 0.0.18
+
+### Patch Changes
+
+- Update stripe create checkout session middleware
+
+## 0.0.17
+
+### Patch Changes
+
+- Update stripe router middleware being unable to capture stripeConfig due to withApiAuth
+
+## 0.0.16
+
+### Patch Changes
+
+- Update stripe package subpath exports
+
+## 0.0.15
+
+### Patch Changes
+
+- Update export subpath in stripe package
+
+## 0.0.14
+
+### Patch Changes
+
+- In Stripe package.json add server exports subpath
+
+## 0.0.13
+
+### Patch Changes
+
+- Update stripeConfig in middleware"
+
+## 0.0.12
+
+### Patch Changes
+
+- Update Stripe package api routes
+
+## 0.0.11
+
+### Patch Changes
+
+- Add stripeConfig.apiRoutes
+
+## 0.0.10
+
+### Patch Changes
+
+- Encapsulate initStripeSupabaseAdmin call in middleware to prevent leak"
+
+## 0.0.9
+
+### Patch Changes
+
+- Update StripeRouterMiddleware routes
+
+## 0.0.8
+
+### Patch Changes
+
+- Add StripeRouterMiddleware
+
+## 0.0.7
+
+### Patch Changes
+
+- Split Stripe package into client and server to prevent imports from being called
+
+## 0.0.6
+
+### Patch Changes
+
+- Add debug info to stripe"
+
 ## 0.0.5
 
 ### Patch Changes